Fix the MAC address of the endpoint port
The endpoint uses its id to overwrite the last byte of the MAC address;
it will basically overwrite the last byte with 0xff. This doesn't make
any sense if the given hardware address should be the base address. All
actual ports will use base+1, base+2 and so on. Therefore, just use
base+0 for the endpoint

Signed-off-by: Michael Walle <michael.walle@kontron.com>